% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr.path}
\alias{rrr.path}
\title{Solution path for reduced rank estimators}
\usage{
rrr.path(Y, X, penaltySVD = c("count", "soft", "ann"), maxrank = min(dim(Y),
  dim(X)), gamma = 2, nlambda = 100, lambda = NULL, control = list())
}
\arguments{
\item{Y}{a matrix of response (n by q)}

\item{X}{a matrix of covariate (n by p)}

\item{penaltySVD}{the penalty to be used on the SVD:
`count' means rank-constrainted estimation;
`soft' means soft singular value thresholding;
`ann' means adaptive soft singular value thresholding.}

\item{maxrank}{an integer of maximum desired rank.}

\item{gamma}{a scalar power parameter of the adaptive weights in penalty == "ann"}

\item{nlambda}{the number of lambda values; no effect if
penalty == `count'.}

\item{lambda}{a vector of user-specified lambda values.}

\item{control}{a list of parameters for controlling the fitting process:
`sv.tol' controls the tolerence of singular values;
`qr.tol' controls the tolerence of QR decomposition for the LS fit}
}
\value{
S3 \code{rrr.path} object, a list consisting of
  \item{A}{right singular matrix}
  \item{Ad}{a vector of singular values}
  \item{coef.ls}{coefficient estimate from LS}
  \item{Spath}{a matrix, each column containing shrinkage factors of the singular values of a solution}
  \item{df.exact}{the exact degrees of freedom}
  \item{df.naive}{the naive degrees of freedom as the effective number of parameters}
  \item{penaltySVD}{the method of low-rank estimation}
}
\description{
Solution path for reduced rank estimators
}
\examples{
Y <- matrix(rnorm(400), 100, 4)
X <- matrix(rnorm(800), 100, 8)
rpath <- rrr.path(Y, X, maxrank = 3)
}

