% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofar_sim.R
\name{simmodel1}
\alias{simmodel1}
\title{Simulation model 1}
\usage{
simmodel1(n = 50, p = 25, q = 25, nrank = 3, s2n = 1, sigma = NULL,
  rho_X = 0.5, rho_E = 0)
}
\arguments{
\item{n, p, q}{model dimensions}

\item{nrank}{model rank}

\item{s2n}{signal to noise ratio}

\item{sigma}{error variance. If specfied, then s2n has no effect}

\item{rho_X}{correlation parameter in the generation of predictors}

\item{rho_E}{correlation parameter in the generation of random errors}
}
\value{
similated model and data
}
\description{
similar to the the simulation model in Chen, Chan, Stenseth (2012), JRSSB
}

