% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofar_sim.R
\name{simmodel2}
\alias{simmodel2}
\title{Simulation model 2}
\usage{
simmodel2(n = 100, p = 50, p0 = 10, q = 50, q0 = 10, nrank = 3,
  s2n = 1, sigma = NULL, rho_X = 0.5, rho_E = 0)
}
\arguments{
\item{n}{sample size}

\item{p}{number of predictors}

\item{p0}{number of relevant predictors}

\item{q}{number of responses}

\item{q0}{number of relevant responses}

\item{nrank}{model rank}

\item{s2n}{signal to noise ratio}

\item{sigma}{error variance. If specfied, then s2n has no effect}

\item{rho_X}{correlation parameter in the generation of predictors}

\item{rho_E}{correlation parameter in the generation of random errors}
}
\value{
similated model and data
}
\description{
similar to the the simulation model in Chen and Huang (2012), JASA
}

