% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofar.R
\name{sofar.control}
\alias{sofar.control}
\title{Internal function for specifying computation parameters}
\usage{
sofar.control(lamA = NULL, lamB = NULL, lamD = NULL,
  methodA = "adlasso", methodB = "adlasso", epsilon = 0.001,
  maxit = 200L, innerEpsilon = 0.001, innerMaxit = 50L, sv.tol = 0.01)
}
\arguments{
\item{lamA}{sequence of tuning parameters for A}

\item{lamB}{sequence of tuning parameters for B}

\item{lamD}{sequence of tuning parameters for d}

\item{methodA}{penalty for penalizing A}

\item{methodB}{penalty for penalizing B}

\item{epsilon}{convergence tolerance}

\item{maxit}{maximum number of iterations}

\item{innerEpsilon}{convergence tolerance for inner subroutines}

\item{innerMaxit}{maximum number of iterations for inner subroutines}

\item{sv.tol}{tolerance for singular values}
}
\value{
a list of computational parameters.
}
\description{
a list of internal computational parameters controlling optimization
}

