% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofar.R
\name{sofar.modstr}
\alias{sofar.modstr}
\title{Internal function for specifying model parameters}
\usage{
sofar.modstr(mu = 1, mugamma = 1.1, WA = NULL, WB = NULL, Wd = NULL,
  wgamma = 2)
}
\arguments{
\item{mu}{parameter in the augmented Lagrangian function}

\item{mugamma}{increament of mu along iterations to speed up computation}

\item{WA}{weight matrix for A}

\item{WB}{weight matrix for B}

\item{Wd}{weight matrix for d}

\item{wgamma}{power parameter in constructing adaptive weights}
}
\value{
a list of model parameters.
}
\description{
a list of internal model parameters controlling the model fitting
}

