% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofar.R
\name{sofar.path}
\alias{sofar}
\alias{sofar.path}
\title{Compute solution paths of sparse orthogonal factor regression}
\usage{
sofar.path(Y, X, nrank = 1, nlam = 50, lam.min.factor = 1e-04,
  lam.max.factor = 10, lam.AB.factor = 1, penA = TRUE, penB = TRUE,
  penD = TRUE, tuning = c("CV", "BIC", "GIC", "AIC", "GCV")[3], nfold = 5,
  norder = NULL, seed = NULL, modstr = list(), control = list())
}
\arguments{
\item{Y}{response matrix}

\item{X}{covariate matrix}

\item{nrank}{an integer specifying the desired rank/number of factors}

\item{nlam}{number of lambda triplets to be used}

\item{lam.min.factor}{set the smallest lambda triplets as a fraction of the estimation lambda.max triplets}

\item{lam.max.factor}{set the largest lambda triplets as a multiple of the estimation lambda.max triplets}

\item{lam.AB.factor}{set the relative penalty level between A/B and D}

\item{penA, penB, penD}{if TRUE, penalty is applied}

\item{tuning}{select tuning method; the default is GIC}

\item{nfold, norder, seed}{used if tuning = "CV"}

\item{modstr}{a list of internal model parameters controlling the model fitting}

\item{control}{a list of internal computation parameters controlling optimization}
}
\value{
\item{U}{estimated U matrix (factor weights)}
  \item{D}{estimated singular values}
  \item{V}{estimated V matrix (factor loadings)}
  \item{Upath}{solution path of U}
  \item{Dpath}{solution path of D}
  \item{Vpath}{solution path of D}
}
\description{
Compute solution paths of sparse orthogonal factor regression
}
\examples{
#require(sofar)

# Simulate data from a sparse factor regression model
p <- 300; q <- 100; n <- 100
nrank <- 3 
mydata <- simmodel1(n=n,p=p,q=q,nrank=nrank,s2n=1,sigma=NULL,rho_X=0.5,rho_E=0.3)
Y <- mydata$Y
X <- mydata$X

fit1 <- sofar.path(Y,X, tuning = ifelse(n>p,"BIC","GIC"), nrank=nrank+2, 
control = list(methodA="adlasso",methodB="adlasso"))
fit1$U
crossprod(fit1$U) #check orthogonality
fit1$V
crossprod(fit1$V) #check orthogonality
}
\references{
Y. Uematsu, K. Chen, Y. Fan, J. Lv, and W. Lin. (2017) \emph{ Large-scale sparse orthogonal factor regression.}
}

